Move yargs-related code to openapi-codegen-cli
