+++ packages/openapi-codegen-cli/src/commands/common.ts
