+++ packages/openapi-codegen-cli/src/commands/docs-json-api.ts
@@ -1,18 +1,17 @@
-import { createContext, Params } from '@fresha/openapi-codegen-utils';
-
-import { Generator } from './parts';
+import {
+  createContext,
+  Generator,
+  CreateContextParams,
+} from '@fresha/openapi-codegen-docs-json-api';
 
 import type { ArgumentsCamelCase } from 'yargs';
-
-export { builder } from '@fresha/openapi-codegen-utils';
 
 export const command = 'docs-json-api';
 
 export const description = 'generate documentation for JSON:API APIs (experimental)';
 
-export const handler = (args: ArgumentsCamelCase<Params>): void => {
+export const handler = (args: ArgumentsCamelCase<CreateContextParams>): void => {
   const context = createContext(args);
   const generator = new Generator(context);
-
   generator.run();
 };
