+++ packages/openapi-codegen-docs/src/context.ts
@@ -1,0 +1,9 @@
+export { createContext } from '@fresha/openapi-codegen-utils';
+
+export interface CreateContextParams {
+  input: string;
+  output: string;
+  jsonApi?: boolean;
+  verbose?: boolean;
+  dryRun?: boolean;
+}
