+++ packages/openapi-codegen-server-elixir/src/index.ts
@@ -1,1 +1,2 @@
 export { Generator } from './parts';
+export { createContext, CreateContextParams } from './context';
