+++ packages/openapi-codegen-server-mock/src/index.ts
@@ -1,1 +1,2 @@
 export { Generator } from './Generator';
+export { createContext, CreateContextParams } from './context';
