+++ packages/openapi-codegen-utils/src/index.ts
@@ -5,4 +5,3 @@ export * from './jsonapi';
 export * from './logging';
 export * from './openapi';
 export * from './string';
-export * from './yargs';
